Optimisation update?
Fixed error with error handling; it kept trapping errors beyond the check for non-existing CO5 files.

Redid the controls input loop based on feedback from spacemoai. Not sure if how I ended up doing it is actually faster; should perhaps benchmark that.
Instead of a whole bunch of IF checks for all possible keys, I use the control key's index in a string containing all possible keys using INSTR, and GOSUB based on that.
Also replaced the STICK(0) reading with the INPUT$(1) I'm already using for the other controls; it feels like movement is smoother as a result?

Added a sprite that should resemble a floppy disk; it will show up now while saving data.

Moved the code for getting the time-based file basename to a subroutine.

Moved the check for whether or not the placement sprite should be shown, to the sprite display subroutine instead.

You can now exit the application by pressing the ESCape key while on the canvas.

Added Z as an alternative to x for decreasing your Z-axis