Redid tileset menu and replace faces with 'roads' and 'town'
Made it easier to add tilesets to the menu, and remove the proof-of-concept faces listing with ones for the 'roads' and 'town' sets.

Note: the 'town' set shows a lot more different tiles than can currently be used, but this is something I hope to address in the near future.