E25: It's a small town, you say?
Released version 25.
Changes:
Fixed error with error handling; it kept trapping errors beyond the check for non-existing CO5 files.

Redid the controls input loop based on feedback from spacemoai. Not sure if how I ended up doing it is actually faster; should perhaps benchmark that.
Instead of a whole bunch of IF checks for all possible keys, I use the control key's index in a string containing all possible keys using INSTR, and GOSUB based on that.
Also replaced the STICK(0) reading with the INPUT$(1) I'm already using for the other controls; it feels like movement is smoother as a result?

Added a sprite that should resemble a floppy disk; it will show up now while saving data.

Moved the code for getting the time-based file basename to a subroutine.

Moved the check for whether or not the placement sprite should be shown, to the sprite display subroutine instead.

You can now exit the application by pressing the ESCape key while on the canvas.

Added Z as an alternative to x for decreasing your Z-axis

Moved files that are necessary for running the programme to their own folder, and moved old releases to a separate folder outside it.
This way you can just mount the dir-as-dsk folder in openMSX, and not have to worry about Git files and other project files taking up space.

Once I've gone through them, I'll also add some of my random script I used to figure out how certain things work, to the testscripts directory.

Made it easier to add tilesets to the menu, and remove the proof-of-concept faces listing with ones for the 'roads' and 'town' sets.

Note: the 'town' set shows a lot more different tiles than can currently be used, but this is something I hope to address in the near future.